app/requirements.txt
discord.py>=2.3.0
asyncpg>=0.28.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
redis>=4.5.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pydantic>=2.0.0
structlog>=23.0.0